+++ src/index.html
@@ -25,15 +25,67 @@
       </section>
     </section>
 
-    <section class="manager-page hide"></section>
 
+    <!-- $('.manager-page').html(` -->
+    <section class="manager-page hide">
+      <section class="main-manager-info">
+        <p class="todays-date"></p>
+        <section class="manager-search-nav">
+          <section class="user-pop-up-window hide"></section>
+          <label class="search-nav-title">Search Customers Name</label>
+          <input class="input-names" type="text"></input>
+          <button id="search-user" type="submit">search</button>
+        </section>
+        <section class="manager-navs">
+          <section class="nav-box">
+            <p class="all-available-rooms-txt">Available Rooms:</p>
+            <p class="all-available-rooms"></p>
+          </section>
+          <section class="nav-box">
+            <p class="total-revenue-text">Total Revenue:</p>
+            <p class="total-revenue"></p>
+          </section>
+          <section class="nav-box">
+            <p class="available-percent-text">Percent of Rooms available:</p>
+            <p class="available-percent"></p>
+          </section>
+        </section>
+      </section>
+    </section>
+
+    <!-- $('.user-page').html(` -->
     <section class="user-page hide">
-        
-
-
-
+      <section class="bookings-container">
+        <section class="bookings-box">
+          <p class="my-bookings-text">My Bookings</p>
+          <section class="my-bookings"></section>
+        </section>
+        <section class="nav-box">
+          <p class=" money-spent-text">Money Spent</p>
+          <section class="money-spent"></section>
+        </section>
+      </section>
+      <section class="main-user-info">
+        <p class="welcome-user"></p>
+        <section class="user-search-nav">
+          <label class="search-nav-title">Check Room Availability</label>
+          <p class="date-helper">date format 'yyyy/mm/dd'</p>
+          <section class="pop-up-container hide"></section>
+          <input class="input-name" type="test"></input>
+          <label class="search-nav-title">Filter By Room-Type</label>
+          <select class="filter-input" type="">
+            <option value="none">none</option>
+            <option>residential suite</option>
+            <option>suite</option>
+            <option>single room</option>
+            <option>junior suite</option>
+          </select>
+          <button id="search-btn" type="submit">search</button>
+        </section>
+      </section>
   </main>
-  <!-- Do not include the index.js or main.bundle.js file here - it is done automatically by the webpack server -->
+  <!-- Do not
+            include the index.js or main.bundle.js file here - it is done automatically by the webpack server -->
 </body>
 
 </html>
